refactor: 인코딩 유틸 및 App 구조 리팩토링
- encodingUtils 함수 개선
- App 컴포넌트 구조 정리
